Widen the retry counter, and let a scheduled event wait where a message may not
Two findings from the seventh review pass, both about a value or a record
outliving what it describes.

`attempts` on the schedule was a `PositiveSmallIntegerField`, and
`--max-attempts 0` retries a row without end -- four months of the default
lease reaches 32767, and PostgreSQL then refuses the increment and fails the
pass rather than the row. It is a `PositiveBigIntegerField` now, so there is no
lease short enough or deployment long enough to reach the end of it.

The same class, one file over: `--max-attempts` takes any positive number, and
a bound above 32767 writes its give-up event into the feed's narrow `attempt`
column -- where a value that does not fit fails the batch of two hundred it
travelled in. `eventlog.writer` saturates it, the way it already truncates text
to its width, and the exact count stays in `detail.attempts`. Widening that
column is a table rewrite of the largest thing this package has, for a number
nobody reads past a handful.

And `after_commit` now tests a weaker condition than `defer` does, which is the
point of the two being separate functions. Under manual transaction management
an `atomic()` block does not commit on exit, so its hooks wait for
`set_autocommit(True)` -- measured: they run after the caller's `commit()` and
not after its `rollback()`. For a publish that is unacceptable, so `defer`
refuses there. For a durable event it is exactly right: arriving late costs a
record nothing, and existing when the row it describes was rolled away costs it
everything. Only autocommit off with no block anywhere is left with no hook to
take, and that is where the event is still recorded at once.

Four cases, each failing without its fix: the manually managed rollback keeps no
event, the manually managed commit still lands one, the retry counter has no
reachable ceiling, and a bound past the feed's column still records the give-up.
The column range is asserted against Django's portable table rather than this
backend -- SQLite enforces no range at all, so a case driving the real column
would have been green with the narrow field in place.

+++ src/django_aiogram/migrations/0004_scheduled_send.py
@@ -42,7 +42,7 @@ class Migration(migrations.Migration):
                 ('claimed_at', models.DateTimeField(blank=True, null=True)),
                 ('claimed_by', models.CharField(blank=True, max_length=128)),
                 ('claimed_until', models.DateTimeField(blank=True, null=True)),
-                ('attempts', models.PositiveSmallIntegerField(default=0)),
+                ('attempts', models.PositiveBigIntegerField(default=0)),
             ],
             options={
                 'db_table': 'django_aiogram_scheduled',
